improvement(webhooks): styling alignment and route refactor and de-deuplication detection
