--- sim/app/api/webhooks/utils.ts
@@ -1,66 +1,0 @@
-/**
- * Validates a Slack webhook request signature using HMAC SHA-256
- * @param signingSecret - Slack signing secret for validation
- * @param signature - X-Slack-Signature header value
- * @param timestamp - X-Slack-Request-Timestamp header value
- * @param body - Raw request body string
- * @returns Whether the signature is valid
- */
-
-export async function validateSlackSignature(
-  signingSecret: string,
-  signature: string,
-  timestamp: string,
-  body: string
-): Promise<boolean> {
-  try {
-    // Basic validation first
-    if (!signingSecret || !signature || !timestamp || !body) {
-      return false
-    }
-
-    // Check if the timestamp is too old (> 5 minutes)
-    const currentTime = Math.floor(Date.now() / 1000)
-    if (Math.abs(currentTime - parseInt(timestamp)) > 300) {
-      return false
-    }
-
-    // Compute the signature
-    const encoder = new TextEncoder()
-    const baseString = `v0:${timestamp}:${body}`
-
-    // Create the HMAC with the signing secret
-    const key = await crypto.subtle.importKey(
-      'raw',
-      encoder.encode(signingSecret),
-      { name: 'HMAC', hash: 'SHA-256' },
-      false,
-      ['sign']
-    )
-
-    const signatureBytes = await crypto.subtle.sign('HMAC', key, encoder.encode(baseString))
-
-    // Convert the signature to hex
-    const signatureHex = Array.from(new Uint8Array(signatureBytes))
-      .map((b) => b.toString(16).padStart(2, '0'))
-      .join('')
-
-    // Prepare the expected signature format
-    const computedSignature = `v0=${signatureHex}`
-
-    // Constant-time comparison to prevent timing attacks
-    if (computedSignature.length !== signature.length) {
-      return false
-    }
-
-    let result = 0
-    for (let i = 0; i < computedSignature.length; i++) {
-      result |= computedSignature.charCodeAt(i) ^ signature.charCodeAt(i)
-    }
-
-    return result === 0
-  } catch (error) {
-    console.error('Error validating Slack signature:', error)
-    return false
-  }
-}
