+++ sim/app/w/[id]/components/workflow-block/components/sub-block/components/webhook/components/ui/webhook-url.tsx
@@ -1,4 +1,5 @@
-import { Check, Copy, Loader2 } from 'lucide-react'
+import { useState } from 'react'
+import { CheckCheck, Copy } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
@@ -17,27 +18,37 @@ export function WebhookUrlField({
   copyToClipboard,
 }: WebhookUrlFieldProps) {
   return (
-    <div className="space-y-2">
-      <Label htmlFor="webhook-url">Webhook URL</Label>
-      <div className="flex items-center space-x-2 pr-1">
-        {isLoadingToken ? (
-          <div className="flex-1 h-10 px-3 py-2 rounded-md border border-input bg-background flex items-center">
-            <Loader2 className="h-4 w-4 animate-spin text-muted-foreground" />
-          </div>
-        ) : (
-          <Input id="webhook-url" value={webhookUrl} readOnly className="flex-1" />
-        )}
+    <div className="space-y-1 mb-4">
+      <Label htmlFor="webhook-url" className="text-sm font-medium">
+        Webhook URL
+      </Label>
+      <div className="flex">
+        <Input
+          id="webhook-url"
+          readOnly
+          value={webhookUrl}
+          className="flex-1 h-10 font-mono text-xs cursor-text"
+          onClick={(e) => (e.target as HTMLInputElement).select()}
+          disabled={isLoadingToken}
+        />
         <Button
           type="button"
-          variant="outline"
           size="icon"
+          variant="outline"
+          className="ml-2 h-10 w-10"
           onClick={() => copyToClipboard(webhookUrl, 'url')}
           disabled={isLoadingToken}
-          className="ml-1"
         >
-          {copied === 'url' ? <Check className="h-4 w-4" /> : <Copy className="h-4 w-4" />}
+          {copied === 'url' ? (
+            <CheckCheck className="h-4 w-4 text-green-500" />
+          ) : (
+            <Copy className="h-4 w-4" />
+          )}
         </Button>
       </div>
+      <p className="text-xs text-muted-foreground mt-1">
+        This is the URL that will receive webhook requests
+      </p>
     </div>
   )
 }
